+++ src/components/drawer/index.ts
@@ -1,0 +1,9 @@
+export { default } from "./Drawer";
+export type {
+  DrawerProps,
+  DrawerPlacement,
+  DrawerSize,
+  DrawerMargin,
+  SpacingOption,
+} from "./Drawer";
+export { DRAWER_MOBILE_SHEET_MEDIA } from "./Drawer";